+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added post-hoc FDP upper bounds for unweighted conformal p-values via
+  `nonconform.fdr.conformal_fdp_upper_bound`.
+
 ## [1.0.1] - 2026-05-20
 
 ### Security